requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
Pillow==8.3.1
django-storages==1.11.1
boto3==1.18.1